requirements.txt
paddleocr==2.8.1
paddlepaddle==2.6.2
opencv-python>=4.10.0
numpy>=1.26.0,<2.0
pandas>=2.2.0
pydantic>=2.7.0
python-dotenv>=1.0.1
Pillow>=10.3.0
gradio>=4.36.0
rapidfuzz>=3.9.0
openai>=1.35.0
boto3>=1.34.0
pytest>=8.2.0
